finished UI
next: finish logic - threading and the like, then maybe add a stop and
pause button. After that port to gtk - it might be worth to use vala to
compile to c and then combine it with my code. This way I'll be able to
optimize the code and avoid shared ownership (Glib::RefPtr).